manager(miuix): relocate Mode + ShortcutType into Miuix screens, drop last util files
