--- manager/app/src/main/java/com/resukisu/resukisu/ui/screen/appprofile/AppProfileUtils.kt
@@ -1,15 +1,0 @@
-package com.resukisu.resukisu.ui.screen.appprofile
-
-import androidx.annotation.StringRes
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.res.stringResource
-import com.resukisu.resukisu.R
-
-enum class Mode(@field:StringRes private val res: Int) {
-    Default(R.string.profile_default),
-    Template(R.string.profile_template),
-    Custom(R.string.profile_custom);
-
-    val text: String
-        @Composable get() = stringResource(res)
-}
